+++ pioreactor/actions/leader/experiment_profile.py
@@ -646,17 +646,11 @@ def execute_experiment_profile(profile_filename: str, experiment: str, dry_run: 
                 # ended early
 
                 # stop all jobs started?
-                if profile.halt_on_early_stop:
-                    with JobManager() as jm:
-                        count_jobs_to_kill = jm.count_jobs(
-                            experiment=experiment, job_source="experiment_profile"
-                        )
-                        logger.info(
-                            f"Exiting profile early will halt {count_jobs_to_kill} jobs started by the profile."
-                        )
-                        jm.kill_jobs(experiment=experiment, job_source="experiment_profile")
+                with JobManager() as jm:
+                    count_jobs_to_kill = jm.count_jobs(experiment=experiment, job_source="experiment_profile")
+                    jm.kill_jobs(experiment=experiment, job_source="experiment_profile")
 
-                logger.notice(f"Exiting profile {profile.experiment_profile_name} early: {len(s.queue)} actions not started.")  # type: ignore
+                logger.notice(f"Exiting profile {profile.experiment_profile_name} early: {len(s.queue)} actions not started, {count_jobs_to_kill} halted.")  # type: ignore
 
             else:
                 if dry_run: